+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Breaking Changes
 * Made all `Configuration` fields final so they can only be initialized in the constructor. This better conveys the immutability of the configuration class. ([#455](https://github.com/realm/realm-dart/pull/455))
+* Removed `inMemory` field from `Configuration`. Use `Configuration.inMemory` factory instead.
+* Due to the introduction of different types of configurations the `Configuration` constructor has been removed. Use the `Configuration.local` factory instead. ([#496](https://github.com/realm/realm-dart/pull/496))
 
 ### Enhancements
 * Added a property `Configuration.disableFormatUpgrade`. When set to `true`, opening a Realm with an older file format will throw an exception to avoid automatically upgrading it. ([#310](https://github.com/realm/realm-dart/pull/310))
@@ -35,6 +37,7 @@
 * Support user deviceId. ([#570](https://github.com/realm/realm-dart/pull/570))
 * Support user authentication provider type. ([#570](https://github.com/realm/realm-dart/pull/570))
 * Support user profile data. ([#570](https://github.com/realm/realm-dart/pull/570))
+* Support flexible synchronization. ([#496](https://github.com/realm/realm-dart/pull/496))
 
 ### Fixed
 * Fixed an issue that would result in the wrong transaction being rolled back if you start a write transaction inside a write transaction. ([#442](https://github.com/realm/realm-dart/issues/442))